+++ src/main/java/ortus/boxlang/modules/orm/mapping/HibernateXMLWriter.java
@@ -253,16 +253,22 @@ public class HibernateXMLWriter {
 	 * @return A &lt;property /&gt; element ready to add to a Hibernate mapping document
 	 */
 	public Element generatePropertyElement( IPropertyMeta prop ) {
-		IStruct	columnInfo	= prop.getColumn();
+		IStruct	propAnnotations	= prop.getAnnotations();
+		IStruct	columnInfo		= prop.getColumn();
 
-		Element	theNode		= this.document.createElement( "property" );
+		Element	theNode			= this.document.createElement( "property" );
 		theNode.setAttribute( "name", prop.getName() );
 		theNode.setAttribute( "type", toHibernateType( prop.getORMType(), false ) );
+		populateStringAttributes( theNode, propAnnotations, List.of( ORMKeys.index ) );
 
 		if ( prop.getFormula() != null ) {
 			theNode.setAttribute( "formula", "( " + prop.getFormula() + " )" );
 		} else {
-			theNode.appendChild( generateColumnElement( prop ) );
+			String[] columnName = columnInfo.getOrDefault( Key._name, "" ).toString().split( "," );
+			for ( String column : columnName ) {
+				columnInfo.put( Key._name, column.trim() );
+				theNode.appendChild( generateColumnElement( columnInfo ) );
+			}
 		}
 		if ( columnInfo.containsKey( ORMKeys.insertable ) ) {
 			theNode.setAttribute( "insert", trueFalseFormat( columnInfo.getAsBoolean( ORMKeys.insertable ) ) );
@@ -312,14 +318,14 @@ public class HibernateXMLWriter {
 		}
 
 		if ( association.containsKey( ORMKeys.inverseJoinColumn ) ) {
-			// @TODO: Loop over all column values and create multiple <column> elements.
-			toManyNode.setAttribute( "column", escapeReservedWords( translateColumnName( association.getAsString( ORMKeys.inverseJoinColumn ) ) ) );
+			addColumnNames( toManyNode, association.getAsString( ORMKeys.inverseJoinColumn ) );
 		}
 
 		if ( association.containsKey( Key._CLASS ) ) {
 			setEntityName( toManyNode, association.getAsString( Key._CLASS ), prop );
 		} else {
-			throw new BoxRuntimeException( "Missing required class name for relationship [%s] on entity".formatted( prop.getName() ) );
+			throw new BoxRuntimeException(
+			    "Missing required class name for relationship '%s' on entity '%s'".formatted( prop.getName(), this.entity.getEntityName() ) );
 		}
 
 		collectionNode.appendChild( toManyNode );
@@ -345,9 +351,6 @@ public class HibernateXMLWriter {
 		if ( type.equals( "map" ) ) {
 			if ( association.containsKey( ORMKeys.structKeyColumn ) ) {
 				Element mapKeyNode = this.document.createElement( "map-key" );
-				theNode.appendChild( mapKeyNode );
-				// Note that Lucee doesn't support comma-delimited values in structKeyColumn
-				mapKeyNode.setAttribute( "column", escapeReservedWords( translateColumnName( association.getAsString( ORMKeys.structKeyColumn ) ) ) );
 				if ( association.containsKey( ORMKeys.structKeyType ) ) {
 					mapKeyNode.setAttribute( "type", association.getAsString( ORMKeys.structKeyType ) );
 				}
@@ -355,19 +358,20 @@ public class HibernateXMLWriter {
 				if ( association.containsKey( ORMKeys.structKeyFormula ) ) {
 					mapKeyNode.setAttribute( "formula", association.getAsString( ORMKeys.structKeyFormula ) );
 				}
+				addColumnNames( mapKeyNode, association.getAsString( ORMKeys.structKeyColumn ) );
+				theNode.appendChild( mapKeyNode );
 			}
 		}
 		if ( association.containsKey( ORMKeys.elementColumn ) ) {
 			Element elementNode = this.document.createElement( "element" );
-			theNode.appendChild( elementNode );
-			// Note that Lucee doesn't support comma-delimited values in elementColumn
-			elementNode.setAttribute( "column", escapeReservedWords( translateColumnName( association.getAsString( ORMKeys.elementColumn ) ) ) );
 			if ( association.containsKey( ORMKeys.elementType ) ) {
 				elementNode.setAttribute( "type", association.getAsString( ORMKeys.elementType ) );
 			}
 			if ( association.containsKey( ORMKeys.elementFormula ) ) {
 				elementNode.setAttribute( "formula", association.getAsString( ORMKeys.elementFormula ) );
 			}
+			addColumnNames( elementNode, association.getAsString( ORMKeys.elementColumn ) );
+			theNode.appendChild( elementNode );
 		}
 
 		theNode.setAttribute( "name", prop.getName() );
@@ -416,29 +420,47 @@ public class HibernateXMLWriter {
 			    )
 			    .findFirst()
 			    .orElseThrow( () -> new BoxRuntimeException(
-			        "Could not find inverse property for association [%s] on entity [%s]".formatted( prop.getName(), this.entity.getEntityName() ) )
+			        "Could not find inverse property for association '%s' on entity '%s'".formatted( prop.getName(), this.entity.getEntityName() ) )
 			    );
 
 			association.put( Key.column, inverseProp.getAssociation().getAsString( Key.column ) );
 		}
 
+		if ( !prop.getCache().isEmpty() ) {
+			theNode.appendChild( generateCacheElement( prop.getCache() ) );
+		}
+
 		// @JoinColumn - https://docs.jboss.org/hibernate/core/3.6/reference/en-US/html/collections.html#collections-foreignkeys
 		if ( association.containsKey( Key.column ) ) {
 			Element keyNode = this.document.createElement( "key" );
-			if ( association.getAsString( Key.column ).split( "," ).length > 1 ) {
-				Stream.of( association.getAsString( Key.column ).split( "," ) ).forEach( column -> {
-					Element columnNode = this.document.createElement( "column" );
-					columnNode.setAttribute( "name", escapeReservedWords( translateColumnName( column.trim() ) ) );
-					keyNode.appendChild( columnNode );
-				} );
-			} else {
-				keyNode.setAttribute( "column", escapeReservedWords( ( association.getAsString( Key.column ) ) ) );
-			}
+			addColumnNames( keyNode, association.getAsString( Key.column ) );
 
 			if ( association.containsKey( ORMKeys.mappedBy ) ) {
 				keyNode.setAttribute( "property-ref", association.getAsString( ORMKeys.mappedBy ) );
 			}
 			theNode.appendChild( keyNode );
+		}
+		return theNode;
+	}
+
+	/**
+	 * Add column names (single, via attribute, or multiple, via child column nodes) to the given node.
+	 * 
+	 * @param theNode    Parent node to either modify or add child nodes to.
+	 * @param columnName The column name(s) to add. If multiple, they should be comma-delimited.
+	 */
+	protected Element addColumnNames( Element theNode, String columnName ) {
+		if ( columnName == null ) {
+			return theNode;
+		}
+		if ( columnName.split( "," ).length > 1 ) {
+			Stream.of( columnName.split( "," ) ).forEach( column -> {
+				Element columnNode = this.document.createElement( "column" );
+				columnNode.setAttribute( "name", escapeReservedWords( translateColumnName( column.trim() ) ) );
+				theNode.appendChild( columnNode );
+			} );
+		} else {
+			theNode.setAttribute( "column", escapeReservedWords( ( columnName ) ) );
 		}
 		return theNode;
 	}
@@ -457,6 +479,9 @@ public class HibernateXMLWriter {
 
 		if ( association.containsKey( Key._CLASS ) ) {
 			setEntityName( theNode, association.getAsString( Key._CLASS ), prop );
+		} else {
+			throw new BoxRuntimeException(
+			    "Missing required class name for relationship '%s' on entity '%s'".formatted( prop.getName(), this.entity.getEntityName() ) );
 		}
 
 		List<Key> stringProperties = List.of( Key._NAME, ORMKeys.cascade, ORMKeys.fetch, ORMKeys.mappedBy, ORMKeys.access,
@@ -474,15 +499,7 @@ public class HibernateXMLWriter {
 			theNode.setAttribute( "formula", prop.getFormula() );
 		}
 		if ( association.containsKey( Key.column ) ) {
-			if ( association.getAsString( Key.column ).split( "," ).length > 1 ) {
-				Stream.of( association.getAsString( Key.column ).split( "," ) ).forEach( column -> {
-					Element columnNode = this.document.createElement( "column" );
-					columnNode.setAttribute( "name", escapeReservedWords( translateColumnName( column.trim() ) ) );
-					theNode.appendChild( columnNode );
-				} );
-			} else {
-				theNode.setAttribute( "column", escapeReservedWords( translateColumnName( association.getAsString( Key.column ) ) ) );
-			}
+			addColumnNames( theNode, association.getAsString( Key.column ) );
 		}
 
 		// for attributes specific to each association type
@@ -500,10 +517,7 @@ public class HibernateXMLWriter {
 				if ( association.containsKey( ORMKeys.unique ) ) {
 					theNode.setAttribute( "unique", trueFalseFormat( association.getAsBoolean( ORMKeys.unique ) ) );
 				}
-				if ( association.containsKey( ORMKeys.missingRowIgnored ) ) {
-					theNode.setAttribute( "not-found", association.getAsString( ORMKeys.missingRowIgnored ) );
-				}
-				// @TODO: unique-key
+				populateStringAttributes( theNode, association, List.of( ORMKeys.uniqueKey, ORMKeys.index ) );
 				break;
 		}
 		return theNode;
@@ -516,13 +530,12 @@ public class HibernateXMLWriter {
 	 *
 	 * @TODO: Refactor all key logic into a getPropertyColumn() method which groups and combines all the various column-specific annotations.
 	 *
-	 * @param prop Column metadata
+	 * @param columnInfo Column metadata
 	 *
 	 * @return A &lt;column /&gt; element ready to add to a Hibernate mapping document
 	 */
-	public Element generateColumnElement( IPropertyMeta prop ) {
+	public Element generateColumnElement( IStruct columnInfo ) {
 		Element		theNode				= this.document.createElement( "column" );
-		IStruct		columnInfo			= prop.getColumn();
 
 		List<Key>	stringProperties	= List.of( Key._DEFAULT, Key.sqltype, ORMKeys.length, ORMKeys.precision, ORMKeys.scale, ORMKeys.uniqueKey );
 		populateStringAttributes( theNode, columnInfo, stringProperties );
@@ -568,7 +581,8 @@ public class HibernateXMLWriter {
 	 * @return An id or key-property XML node ready to add to a Hibernate mapping class or composite-id element
 	 */
 	public Element generateIdElement( String elementName, IPropertyMeta prop ) {
-		Element theNode = this.document.createElement( elementName );
+		Element	theNode		= this.document.createElement( elementName );
+		IStruct	columnInfo	= prop.getColumn();
 
 		// set common attributes
 		theNode.setAttribute( "name", prop.getName() );
@@ -577,11 +591,15 @@ public class HibernateXMLWriter {
 			theNode.setAttribute( "unsaved-value", prop.getUnsavedValue() );
 		}
 
-		theNode.appendChild( generateColumnElement( prop ) );
+		String[] columnName = columnInfo.getOrDefault( Key._name, "" ).toString().split( "," );
+		for ( String column : columnName ) {
+			columnInfo.put( Key._name, column.trim() );
+			theNode.appendChild( generateColumnElement( columnInfo ) );
+		}
 
 		if ( !prop.getGenerator().isEmpty() ) {
 			if ( elementName.equals( "key-property" ) ) {
-				logger.error( "Composite ID elements do not support generators. Ignoring generator for property [%s] on entity [%s].",
+				logger.error( "Composite ID elements do not support generators. Ignoring generator for property [{}] on entity [{}].",
 				    prop.getName(), entity.getEntityName() );
 			} else {
 				theNode.appendChild( generateGeneratorElement( prop.getGenerator() ) );
@@ -898,7 +916,7 @@ public class HibernateXMLWriter {
 		theNode.setAttribute( "type", toHibernateType( prop.getORMType(), false ) );
 		// COLUMN name
 		if ( columnInfo.containsKey( Key._NAME ) ) {
-			theNode.setAttribute( "column", escapeReservedWords( translateColumnName( columnInfo.getAsString( Key._NAME ) ) ) );
+			addColumnNames( theNode, columnInfo.getAsString( Key._NAME ) );
 		}
 		if ( prop.getUnsavedValue() != null ) {
 			theNode.setAttribute( "unsaved-value", prop.getUnsavedValue() );
@@ -937,7 +955,8 @@ public class HibernateXMLWriter {
 	 */
 	private void setEntityName( Element theNode, String relationClassName, IPropertyMeta prop ) {
 		if ( relationClassName == null || relationClassName.isBlank() ) {
-			return;
+			throw new BoxRuntimeException(
+			    "Missing required class name for relationship '%s' on entity '%s'".formatted( prop.getName(), this.entity.getEntityName() ) );
 		}
 		Key				datasourceName		= this.entity.getDatasource().isEmpty() ? Key.defaultDatasource : Key.of( this.entity.getDatasource() );
 		EntityRecord	associatedEntity	= entityLookup.apply( relationClassName, datasourceName );
@@ -950,6 +969,17 @@ public class HibernateXMLWriter {
 				logger.error( message );
 			}
 		} else {
+			if ( associatedEntity.getEntityName() == null || associatedEntity.getEntityName().isBlank() ) {
+				String message = String.format( "Associated entity '%s' referenced in property '%s' on entity '%s' has null or empty entity name",
+				    relationClassName,
+				    prop.getName(),
+				    prop.getDefiningEntity().getEntityName() );
+				if ( !this.ormConfig.ignoreParseErrors ) {
+					throw new BoxRuntimeException( message );
+				} else {
+					logger.error( message );
+				}
+			}
 			theNode.setAttribute( "entity-name", associatedEntity.getEntityName() );
 		}
 	}
@@ -1013,6 +1043,8 @@ public class HibernateXMLWriter {
 				return "order-by";
 			case "uniquekey" :
 				return "unique-key";
+			case "missingRowIgnored" :
+				return "not-found";
 			default :
 				return name;
 		}